added cars create/delete/index

+++ database/migrations/2020_06_26_151557_create_car_marks_table.php
@@ -10,7 +10,7 @@ class CreateCarMarksTable extends Migration
     {
         Schema::create('car_marks', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('country_id');
+            $table->unsignedTinyInteger('country_id');
             $table->string('name');
             $table->string('mark_image_path');
         });
